Move write service state out of tester struct
