+++ internal/testrunner/runners/system/state.go
@@ -10,8 +10,10 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/elastic/elastic-package/internal/agentdeployer"
 	"github.com/elastic/elastic-package/internal/kibana"
 	"github.com/elastic/elastic-package/internal/logger"
+	"github.com/elastic/elastic-package/internal/servicedeployer"
 	"github.com/elastic/elastic-package/internal/stack"
 )
 
@@ -49,4 +51,38 @@ func readServiceStateData(path string) (ServiceState, error) {
 		return setupData, fmt.Errorf("failed to decode service options %q: %w", path, err)
 	}
 	return setupData, nil
+}
+
+type scenarioStateOpts struct {
+	currentPolicy *kibana.Policy
+	enrollPolicy  *kibana.Policy
+	origPolicy    *kibana.Policy
+	config        *testConfig
+	agent         kibana.Agent
+	agentInfo     agentdeployer.AgentInfo
+	svcInfo       servicedeployer.ServiceInfo
+}
+
+func writeScenarioState(opts scenarioStateOpts, target string) error {
+	data := ServiceState{
+		OrigPolicy:       *opts.origPolicy,
+		EnrollPolicy:     *opts.enrollPolicy,
+		CurrentPolicy:    *opts.currentPolicy,
+		Agent:            opts.agent,
+		ConfigFilePath:   opts.config.Path,
+		VariantName:      opts.config.ServiceVariantName,
+		ServiceRunID:     opts.svcInfo.Test.RunID,
+		AgentRunID:       opts.agentInfo.Test.RunID,
+		ServiceOutputDir: opts.svcInfo.OutputDir,
+	}
+	dataBytes, err := json.Marshal(data)
+	if err != nil {
+		return fmt.Errorf("failed to marshall service setup data: %w", err)
+	}
+
+	err = os.WriteFile(target, dataBytes, 0644)
+	if err != nil {
+		return fmt.Errorf("failed to write service setup JSON: %w", err)
+	}
+	return nil
 }
